mcp_agent/playground/retriever/requirements.txt
faiss-cpu>=1.7.4
langchain-community>=0.0.28
langchain-openai>=0.1.1
openai>=1.12.0
python-dotenv>=1.0.0
numpy>=2.0.0
